Fixed issues with stack handling and error handling.
Fixed error where the operator stack wasn't reducing in value, or
removing the operator from the stack properly.  This caused an errror
when trying to convert the equation to RPN if the higher priority
operator was on the left.
Added two more tests to the test/index.js to confirm this is now working
(previous tests were not extensive enough, and missed this.)
Also modified the error messaging within the parser, since errors were
being missed due to the stack issue, and they weren't working as
intended.